All changes from GSoC 2023

+++ nwb_table_readme.py
@@ -7,21 +7,40 @@ import math
 import datalad.api as dl
 import json
 import argparse
+import tempfile
+
 from datetime import date
 from pynwb import NWBHDF5IO
 from pynwb.image import ImageSeries
 from pynwb.base import TimeSeries
 from pynwb.behavior import BehavioralTimeSeries, BehavioralEvents
 from dandi.pynwb_utils import get_nwb_version
-from nwbinspector import inspect_nwb
+from nwbinspector import inspect_nwbfile
 from nwbinspector.register_checks import Importance
 from nwbinspector.inspector_tools import save_report, format_messages, MessageFormatter
 from dandi import download
 import ast
 from collections import Counter
+from contextlib import contextmanager
+from create_summary import create_summary
+
+
+
+class TimeoutException(Exception): pass
+
+@contextmanager
+def time_limit(seconds):
+    def signal_handler(signum, frame):
+        raise TimeoutException("Timed out!")
+    signal.signal(signal.SIGALRM, signal_handler)
+    signal.alarm(seconds)
+    try:
+        yield
+    finally:
+        signal.alarm(0)
 
 def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dandisetlimit=None,
-                            args_updatereadme=None,args_readmeonly=None):
+                            args_updatereadme=None,args_readmeonly=None,args_testdocker=False,args_createsummary=None):
     # if users want to update readme only
     if args_readmeonly:
         return args_readmeonly
@@ -33,23 +52,27 @@ def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dand
     json_file = '.dandi/assets.json'
     # directory for dandisets
     root_folder = '/tmp/dandisets'
-    if os.path.exists(root_folder):
-        if len(os.listdir((root_folder))) != 0:
-            dl.update(how='merge', how_subds='reset', follow='parentds-lazy', recursive=True)
+    if args_testdocker is False:
+        if os.path.exists(root_folder):
+            if len(os.listdir((root_folder))) != 0:
+                dl.update(how='merge', how_subds='reset', follow='parentds-lazy', recursive=True)
+            else:
+                os.mkdir(root_folder)
+                dl.install(source='https://github.com/dandi/dandisets.git', path=root_folder, recursive=True, recursion_limit=1, jobs=4)
         else:
-            os.mkdir(root_folder)
             dl.install(source='https://github.com/dandi/dandisets.git', path=root_folder, recursive=True, recursion_limit=1, jobs=4)
-    else:
 
-        print("Fresh install of dandisets...")
-        dl.install(source='https://github.com/dandi/dandisets.git', path=root_folder, recursive=True, recursion_limit=1, jobs=4)
-
-    print("Completed install of dandisets...")
     # directory for storing validation files and readme file
-    save_folder = 'validation_folder'
+    if(args_testdocker):
+        save_folder = 'testing/validation_folder'
+        summary_folder = os.path.join(save_folder,'Summaries')
+    else:
+        save_folder = 'validation_folder'
+        summary_folder = 'Summaries'
+        
     if not os.path.exists(save_folder):
         os.mkdir(save_folder)
-
+        
     dandiset_folder_name = sorted([item for item in os.listdir(root_folder) if item.startswith('0')])
 
     # if user only wants to run the script for 10 dandisets
@@ -64,7 +87,7 @@ def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dand
 
     dandi_metadata = pd.DataFrame()
     nanval = math.nan
-
+     
     for dandiset_name in dandiset_folder_name:
         print("\n     =================  Dealing with DANDISET ID: %s" % dandiset_name)
         with open(os.path.join(root_folder,dandiset_name,yaml_file)) as f:
@@ -82,17 +105,18 @@ def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dand
             doi_link = my_dict['relatedResource'][0]['url']
         except:
             doi_link = nanval
-
         # flatten the yaml file and read in into dataframe
         yaml_df = pd.json_normalize(my_dict)
 
         # get nwb version for NWB dandisets, dummy for datasets that are not NWB
         nwb_version = nanval
+        html_info = [nanval,nanval]
         smallest_size_lst = [nanval,nanval]
         validation_summary = nanval
         url_lst = [nanval, nanval]
         file_parent_folder = [nanval,nanval]
         nwbe_compatibility = ['NI','NI']
+        
         if data_type != nanval and 'NWB' in str(data_type):
             # get the json file that has individual files info
             with open(os.path.join(root_folder, dandiset_name, json_file)) as f:
@@ -148,7 +172,6 @@ def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dand
                         file_parent_folder.append(nanval)
                         path_lst.append(nanval)
                         break
-
             report_message = []
             # if user doesn't want to download files
             if args_nodownload:
@@ -159,53 +182,72 @@ def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dand
                 validation_summary = 'NULL_FILE_LIMIT'
                 for i in range(len(path_lst)):
                     if smallest_size_lst[i] < hard_limit:
+                    
                         # download files
                         nwb_path = download_nwb_with_path(url_lst[i],path_lst[i])
+                        
                         # get nwb_version
-                        nwb_version = get_nwb_version(nwb_path)
+                        nwb_version = get_nwb_version(nwb_path)                                 
                         # validate file here first
                         try:
-                            report_message.extend(list(inspect_nwb(nwbfile_path=nwb_path,
+                            report_message.extend(list(inspect_nwbfile(nwbfile_path=nwb_path,
                                                                    importance_threshold=Importance.BEST_PRACTICE_VIOLATION)))
-
-                            validation_summary = nwb_inspector_message_format(report_message, dandiset_name, save_folder)
+                            validation_summary = nwb_inspector_message_format(report_message, dandiset_name, save_folder)      
                         except ValueError:
                             validation_summary = 'UNABLE'
-                            pass
+                            continue
                         # test nwbe compatibility
-                        nwbe_compatibility[i] = test_nwbe_compatibility(nwb_path)
+                        if args_createsummary and not os.path.exists(os.path.join(os.path.join(summary_folder,dandiset_name),'file_'+str(i)+'/README.md')):
+                            try:
+                                with time_limit(30):
+                                    print("Creating Summary !")
+                                    create_summary(nwb_path,dandiset_name,str(i))
+                            except:
+                                print("Summary Timeout !")
+                                if os.path.exists(os.path.join(os.path.join(summary_folder,dandiset_name),'file_'+str(i)+'/README.md')):
+                                    os.remove(os.path.join(os.path.join(summary_folder,dandiset_name),'file_'+str(i)+'/README.md'))
+                                pass
+                        nwbe_compatibility[i] = test_nwbe_compatibility(nwb_path,args_testdocker)
                         # uninstall file
                         os.unlink(nwb_path)
+                        
         # concatenate the additional variables to the flattened pdf
         yaml_df = pd.concat([yaml_df, pd.DataFrame([[species_name,data_type,doi_link,nwb_version,validation_summary,
                                                        smallest_size_lst[0],smallest_size_lst[1],url_lst[0],url_lst[1],nwbe_compatibility[0],nwbe_compatibility[1],
                                                      file_parent_folder[0],file_parent_folder[1]]],
                                                    index=yaml_df.index,columns=tmp_col)],axis=1)
-
         # concatenate every newly read dandiset metadata dataframe
         dandi_metadata = pd.concat([dandi_metadata,yaml_df],axis=0,ignore_index=True)
         # in case script crashes
         dandi_metadata.to_csv(os.path.join(save_folder,'dandiset_summary_tmp.csv'))
         f.close()
 
+
     # only get the relevant columns
     yaml_df_flatten.extend(tmp_col)
     dandi_metadata_final = dandi_metadata[yaml_df_flatten].sort_values(by=['identifier'],ignore_index=True)
+    dandi_metadata_final.to_csv(os.path.join(save_folder, 'dandiset_summary.csv')) 
     dandi_metadata_final.rename(columns={'assetsSummary.numberOfBytes':'num_bytes','assetsSummary.numberOfFiles':'num_files','assetsSummary.numberOfSubjects':'numb_subjects',
                                          'assetsSummary.variableMeasured':'variableMeasured', 'schemaVersion':'dandiset_schemaver'},inplace=True)
+    
     # filter column through
     dandi_metadata_final.drop(dandi_metadata_final.filter(regex="Unnamed"), axis=1, inplace=True)
     # save table to csv
     dandi_metadata_final.to_csv(os.path.join(save_folder, 'dandiset_summary.csv'))
 
     # remove the cloned dandisets folder
-    dl.remove(dataset=root_folder)
+    #dl.remove(dataset=root_folder)
     return args_updatereadme
 
-def test_nwbe_compatibility(nwb_path):
-    cmd = 'python compatibility_test.py ' + nwb_path  # the external command to run
+def test_nwbe_compatibility(nwb_path,testdocker):
+    if(testdocker):		
+    	cmd = 'docker exec -i nwbe /bin/sh -c \'python -u testing/compatibility_test.py ' + nwb_path + ' --test_docker\''
+    else:
+    	cmd = 'docker exec -i nwbe /bin/sh -c \'python -u testing/compatibility_test.py ' + nwb_path + '\''
+    	
     timeout_s = 60  # how many seconds to wait
     type_hierarchy = set([ImageSeries,TimeSeries,BehavioralTimeSeries,BehavioralEvents])
+    
     # NC-0: file cannot be opened
     try:
         io = NWBHDF5IO(nwb_path,mode='r',load_namespaces=True)
@@ -214,6 +256,9 @@ def test_nwbe_compatibility(nwb_path):
         print('File cannot be opened - NC lvl 0')
         nwbe_compatibility = 'NC-0'
         return nwbe_compatibility
+        
+        
+    
     # dummy var in case file passes NC-1 and is not a TimeSeries type
     nwbe_compatibility = 'LL-V'
     plottable = 0
@@ -238,18 +283,34 @@ def test_nwbe_compatibility(nwb_path):
                     plottable = 1
                     break
     # NC-1: timeout while creating geppetto model
-    try:
-        p = subprocess.Popen([cmd], start_new_session=True, shell=True)
-        p.wait(timeout=timeout_s)
-    except subprocess.TimeoutExpired:
-        print(f'Timeout for {cmd} ({timeout_s}s) expired')
-        os.killpg(os.getpgid(p.pid), signal.SIGTERM)
-        nwbe_compatibility = 'NC-1'
+
+    if(testdocker):
+        with tempfile.TemporaryFile() as tempf:
+                p = subprocess.Popen([cmd], start_new_session=True, shell=True, stdout=tempf)
+                p.wait()
+                tempf.seek(0)
+                text = tempf.read()
+                string_data = text.decode('utf-8')
+                lower_text = string_data.lower()
+                print(lower_text) #removefinal
+        if "failed" in lower_text:
+            nwbe_compatibility = 'NC-1'
+    else:
+        try:
+            p = subprocess.Popen([cmd],start_new_session=True, shell=True)
+            p.wait(timeout=timeout_s)
+        except subprocess.TimeoutExpired:
+            print(f'Timeout for {cmd} ({timeout_s}s) expired')
+            os.killpg(os.getpgid(p.pid),signal.SIGTERM)
+            nwbe_compatibility = 'NC-1'
+            
+    print(nwbe_compatibility)
     return nwbe_compatibility
 
 def download_nwb_with_path(dandi_url,nwb_file_name):
     # create save folder
     save_folder = '/tmp/nwb_versions'
+    
     if not os.path.exists(save_folder):
         os.mkdir(save_folder)
     if '/' in nwb_file_name:
@@ -290,18 +351,27 @@ def nwb_inspector_message_format(report_message,dds_id,save_folder,detailed_repo
         validation_summary = 'PASSED_VALIDATION'
     return validation_summary
 
-def update_readme():
-    save_folder = 'validation_folder'
+def update_readme(testdocker=None):
+    if(testdocker): 
+        save_folder = 'testing/validation_folder'
+        summary_folder = os.path.join(save_folder,'Summaries')
+    else:
+        save_folder = 'validation_folder'
+        summary_folder = 'Summaries'
+        
     rd_file = os.path.join(save_folder,'README.md')
     summary_file = os.path.join(save_folder, 'dandiset_summary.csv')
+    
+    print(summary_file)
+    
     if not os.path.exists(summary_file):
         exit()
     # Getting Datetime from timestamp
+    
     date_time = date.today()
     dandi_metadata_readme = pd.read_csv(summary_file)
     dandi_metadata_readme.drop(dandi_metadata_readme.filter(regex="Unnamed"), axis=1, inplace=True)
     dandi_metadata_readme.to_csv(summary_file, index=False)
-    print(dandi_metadata_readme)
     # summary statistics here
     data_type_dict = dandi_metadata_readme['data_type'].value_counts().to_dict()
     # get data_type values
@@ -358,6 +428,7 @@ def update_readme():
     readme += '\n'
     readme += '- NWB dandisets that pass NWBInspector and thus are possibly NWBE compatible: '
     root_url = 'https://dandiarchive.org/dandiset/'
+    readme = readme[:-2]+'\n\n'
     for ds in pass_nwbinspector:
         readme += '[%s](%s%s), '%(ds, root_url, ds)
     readme = readme[:-2]+'\n\n'
@@ -424,12 +495,17 @@ def update_readme():
                             dandi_link = dandi_metadata_readme['url'].iloc[row] + '/files?location=' + dandi_metadata_readme['parent_folder_' + str(i)].iloc[row] +'%2F'
                         else:
                             dandi_link = dandi_metadata_readme['url'].iloc[row] + '/files?location='
+                        
                         info_link = dandi_metadata_readme['file_' + str(i)].iloc[row].split('/download')[0]
                         file_size = dandi_metadata_readme['file_size_' + str(i)].iloc[row]
                         readme += 'Size: %s MB | \n' % (str(round(int(file_size)/1000000,2)))
                         readme += '[File info](%s) | \n' % (info_link)
                         readme += '[View on DANDI Web](%s) | \n' % (dandi_link)
-                        readme += '[View on NWB Explorer](%s) \n' % (nwbe_link)
+                        readme += '[View on NWB Explorer](%s) ' % (nwbe_link)
+                        if os.path.exists(os.path.join(os.path.join(summary_folder,ref),'file_'+str(i)+'/README.md')):
+                            readme += '| \n[File Summary](%s) \n' % (os.path.join(os.path.join('Summaries',ref),'file_'+str(i)+'/README.md'))
+                        else:
+                            readme +='\n'
 
             else:
                 readme += '- ![#dd0000](https://via.placeholder.com/15/dd0000/dd0000.png) Validation results summary: ' + dandi_metadata_readme['validation_summary'].iloc[row] + '\n\n'
@@ -454,9 +530,13 @@ if __name__ == '__main__':
                         help='update readme file after summary file is created')
     parser.add_argument('--update_readme_only', default=False, action='store_true',
                         help='update readme file without creating summary file')
+    parser.add_argument('--test_docker', default=False, action='store_true',
+                        help='test using the NWBE docker container')
+    parser.add_argument('--create_summary', default=False, action='store_true',
+                        help='Create summaries for files which do not contain a summary README.md')
     args = parser.parse_args()
     update_readme_option=create_dandiset_summary(args.no_download,args.no_sizelimit,args.dandiset_limit,
-                                                 args.update_readme_option,args.update_readme_only)
+                                                 args.update_readme_option,args.update_readme_only,args.test_docker,args.create_summary)
 
     if update_readme_option:
-        update_readme()
+        update_readme(args.test_docker)
